In puavo-bootserver-sync-images.ng, delete files only on demand.
The main benefit of this is that old images and diffs could be
used to patch new images, so it makes sense to delay their
removal until we really need the space.

The algorithm here does not try to find optimal path to get
an image from source images and diffs.

Also some small logging improvements.