In puavo-ers-applet, check that we are running the expected exams a bit more carefully.
We first check if we have session state, and we can
load new exam set even in case the running exams match
the expected ones.  This is because there might be two
sessions with the same set, yet those sessions need to be
kept separate because their answer sets should be distinct.
In case session state is not known, compare the exam UUIDs
with the actually running exams.