Remove the login keyring at login time.
This might be encrypted with a really old password and can not
be unlocked during login.  Normally this should not matter,
but recent versions of Google Chrome want to open this.
Remove this but make a backup once, just in case.

Note that login.keyring file gets regenerated on every login.
Chrome should be happen because it can be unlocked now.
But we should later remove the pam_gnome_keyring.so from
PAM configurations.  However, I do not want to do this now,
but will apply that change for later versions (Trixie),
to see if there are any ill effects from that change.