Restrict primary user sudo permissions.
Changing "puavo.admin.superusers" choices: "primary-user"
is no longer a valid option, but is split into
"limited-primary-user" and "primary-user-is-super".
The "primary-user-is-super" choice provides the old
semantics, meaning that primary user can use his/her
own password for running commands as root.

The new option "limited-primary-user" will be the default,
and provides the necessary permissions for normal computer
stuff.

This is unfortunate, but we have requests for making this
so and we also need to enable this globally because
of the security requirements of the examination mode.