Keep the sudo-permissions default behaviour for a while.
That is, primary users still have sudo-rights (with their passwords),
for a while.  This is to allow (after updating to this change)
setting host puavo-conf "puavo.admin.superusers" to
"primary-user-is-super" where needed without breaking anything,
and later once we change the default to "limited-primary-user"
this will not affect those hosts where more permissions are needed.

I expect to change the default to "limited-primary-user" during
March 2025.